requirements.txt
pyrofork==2.3.43
tgcrypto-pyrofork
pymongo
python-dotenv